depuracion

+++ FUNCP.cc
@@ -188,7 +188,7 @@ int buscar_listen_repetido(t_direccion *tsap_escucha, t_direccion *tsap_remota){
 
 list<buf_pkt, shm_Allocator<buf_pkt> >::iterator buscar_buffer_libre(){
     list<buf_pkt, shm_Allocator<buf_pkt> >::iterator it_libres;
-    printf(stderr,"\nbuscar_buffer_libre: tamanho %d",KERNEL->buffers_libres.size());
+    printf(stderr,"\nbuscar_buffer_libre: tamanho %d",(int)KERNEL->buffers_libres.size());
     if(KERNEL->buffers_libres.empty()) {
         fprintf(stderr,"\nla lista de libres esta vacia-> creación de un buf_pkt");
         buf_pkt nodo;
